+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes to Calva.
 - [Make code more robust in case Jack-in task fails](https://github.com/BetterThanTomorrow/calva/issues/367)
 - [Support Jack-in without file open for single-rooted workspace](https://github.com/BetterThanTomorrow/calva/issues/366)
 - [Show argument list of fn](https://github.com/BetterThanTomorrow/calva/issues/238)
+- [Warnings/Errors should be made more "prominent" in the "Calva says" channel](https://github.com/BetterThanTomorrow/calva/issues/356)
 
 ## [2.0.44] - 05.10.2019
 - [Support for custom project/workflow commands](https://github.com/BetterThanTomorrow/calva/issues/281)